mv Cell to base/cell.py, add join_cells() there
Moves computations of start indices from domain.__init__ in there

update topos/__init__.py
