+++ topos/__init__.py
@@ -1,2 +1,1 @@
-from .base import *
 from .core import *
