+++ topos/base/__init__.py
@@ -1,3 +1,4 @@
 from .set import Set, Iterable, Map, Mapping
 from .hypergraph import Hypergraph, Chain
-from .shape import Shape, Cell
+from .shape import Shape
+from .cell import Cell, join_cells
